web/requirements.txt
Flask
gspread
python-dotenv